feat(guild): guild war, prestige-shop buffs, residence, and leveling
Builds out the guild (Expedition) system on top of client_version/zone-10.0.2_r575:
declare-war handshake, in-world PvP (enemy recolor + attackability + kill
registration), a live kill scoreboard, per-guild win/lost banners plus a
server-wide result announcement, post-war protection, guild residence
(place/demolish/tax), guild prestige-shop buffs, and guild leveling/experience.
Also fixes several previously-unregistered but fully-implemented C2G/G2C guild
packets, corrects three real client opcodes (SCExpeditionWarStatePacket,
SCExpeditionRejoinFailPacket, plus the new SCExpeditionWarDeclarationMoney),
and closes a gap where the guild-invite flow never checked the inviter's and
invitee's top-level faction (Nuia/Haranya/Pirate) matched.

- ExpeditionManager: DeclareWar/EndWar handshake, RegisterWarKill, kill-score
  tracking (Expedition.WarKillsByMember), post-war protection (defender-only,
  1h), Invite's new same-faction check, TryChangeContributionPoints already
  reused by guild buffs/residence refunds.
- BaseUnit/Character/CharacterCombat: war enemies are mutually hostile and
  attackable everywhere (including normally-protected zones), a war-kill is
  registered on PvP death, and a war participant's client re-learns both the
  guild nameplate tag and the enemy war-state tag whenever a unit becomes
  visible (previously blank until the next membership/war-state change).
- New SC/CS packets for the declare-war money confirm step, the kill
  scoreboard, and the win/lose/draw result notification, plus corrected
  opcodes for SCExpeditionWarStatePacket/SCExpeditionRejoinFailPacket - see
  each packet's own doc comment for the Ghidra/frida trail.
- GameNetwork: registers CSExpeditionBuffPacket/CSExpeditionBuffGradePacket/
  CSExpeditionExpAddPacket/CSExpeditionInterestUpatePacket/
  CSRequestDeclarationMoneyPacket/CSExpeditionLevelUpPacket and a few other
  guild C2G packets that were fully implemented but never wired up.
- HousingManager/HousingGameData: Guild Residence placement (one per guild,
  any member may place/demolish), tax-info push on placement and login, and
  an 80%-of-shop-price Contribution Point refund on demolish.
- Features.json: flips on the client feature bits gating the guild info
  panel, guild-war UI, guild summon, and item-based guild rename - all of
  which have working server support already but were never advertised.
- GuildExp/GuildPrestige/GuildResidence/GuildWarEnd/GuildWarTime/
  EndGuildProtection: GM commands used to build and verify this feature.

Every change here was tested live against a running server/client. Not
included: anything related to the Hero-election system, castle/Dominion
territory ownership, or the login server - those are separate, unfinished
systems this branch does not touch.

+++ AAEmu.Game/Core/Packets/G2C/SCExpeditionOwnerChangedPacket.cs
@@ -3,13 +3,19 @@ using AAEmu.Game.Core.Network.Game;
 
 namespace AAEmu.Game.Core.Packets.G2C;
 
+/// <summary>
+/// 2026-09-02: both id fields widened to 8 bytes - confirmed against the real client
+/// (x2game.dll: PacketFunctor&lt;...,SCExpeditionOwnerChangedPacket&gt; -&gt; FUN_3933a7d0), which reads
+/// two 8-byte values before the name string starts. The previous 4-byte writes would have under-filled
+/// the body by 8 bytes total, misaligning the trailing name string.
+/// </summary>
 public class SCExpeditionOwnerChangedPacket(uint id, uint id2, string charName)
     : GamePacket(SCOffsets.SCExpeditionOwnerChangedPacket, 1)
 {
     public override PacketStream Write(PacketStream stream)
     {
-        stream.Write(id);
-        stream.Write(id2);
+        stream.Write((ulong)id);
+        stream.Write((ulong)id2);
         stream.Write(charName);
         return stream;
     }
